requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
# Install a CUDA-enabled PyTorch/torchvision wheel on Linux first if the host does not already provide one.
torch>=2.6.0
torchvision>=0.21.0
timm>=1.0.11
numpy>=1.26.4
Pillow>=10.0.0
tqdm>=4.66.0